requirements.txt
aldryn-apphooks-config==0.4.1
aldryn-boilerplates==0.7.7
aldryn-bootstrap3==1.3.0
aldryn-categories==1.1.0
aldryn-common==1.0.4
aldryn-events==3.0.0
aldryn-faq==2.1.2
aldryn-forms==3.0.4
aldryn-jobs==3.0.0
aldryn-newsblog==2.1.1
aldryn-people==2.1.0
aldryn-search==1.0.1
aldryn-translation-tools==0.2.1
aldryn-wow==1.0.0
Babel==2.6.0
backport-collections==0.1
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
cmsplugin-filer==1.1.3
cmsplugin-svg==0.1.2
cssselect==1.0.3
cssutils==1.0.2
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11.15
django-absolute==0.3
django-admin-sortable==2.1.7
django-admin-sortable2==0.6.19
django-allauth==0.35.0
django-angular==2.2
django-appconf==1.0.2
django-appdata==0.2.0
django-bootstrap3==11.0.0
django-classy-tags==0.8.0
django-cms==3.5.2
django-compressor==2.2
django-emailit==0.2.3
django-extended-choices==1.3.1
django-extra-views==0.11.0
django-filer==1.3.2
django-filter==1.1.0
django-formtools==2.1
django-fsm==2.6.0
django-fsm-admin==1.2.4
django-haystack==2.5.0
django-ipware==1.1.1
django-meta==1.4.1
django-mptt==0.8.7
django-multiupload==0.5.2
django-oscar==1.6.6
django-parler==1.8
django-phonenumber-field==2.0.1
django-polymorphic==1.3
django-post-office==3.0.3
django-ranged-response==0.2.0
django-redis-cache==1.7.1
django-redis-sessions==0.6.1
django-rest-auth==0.9.2
django-sass-processor==0.6
django-sekizai==0.10.0
django-select2==6.3.1
django-shop==0.12.3
django-simple-captcha==0.5.9
django-sizefield==0.9.1
django-sortedm2m==1.5.0
django-spurl==0.6.5
django-standard-form==1.1.1
django-tables2==1.21.2
django-tablib==3.2
django-taggit==0.23.0
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangocms-admin-style==1.2.8
djangocms-attributes-field==0.3.0
djangocms-bootstrap3==0.3.2
djangocms-cascade==0.16
djangocms-column==1.8.0
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-installer==1.0.1
djangocms-light-gallery==0.0.7
djangocms-link==2.1.2
djangocms-oscar==0.1
djangocms-page-meta==0.8.3
djangocms-picture==2.0.6
djangocms-snippet==2.0.0
djangocms-style==2.0.2
djangocms-text-ckeditor==3.6.0
djangocms-video==2.0.4
djangorestframework==3.7.7
djangoshop-stripe==0.12.3
drf-haystack==1.6.1
easy-thumbnails==2.5
elasticsearch==1.7.0
et-xmlfile==1.0.1
factory-boy==2.11.1
Faker==1.0.1
future==0.16.0
hiredis==0.2.0
html5lib==0.9999999
idna==2.6
jdcal==1.4
jsonfield==1.0.3
libsass==0.12.3
lxml==4.2.5
mock==2.0.0
oauthlib==1.0.3
odfpy==1.3.6
olefile==0.46
openpyxl==2.4.9
pbr==5.1.1
phonenumbers==8.9.13
phonenumberslite==8.9.13
Pillow==4.3.0
premailer==3.2.0
purl==1.4
python-dateutil==2.6.0
python-openid==2.2.5
python-slugify==1.2.0
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
rcssmin==1.0.6
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.5.0
rjsmin==1.0.12
schedule==0.5.0
six==1.11.0
sorl-thumbnail==12.4.1
stripe==1.37.0
tablib==0.12.1
text-unidecode==1.2
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.22
URLObject==2.4.3
virtualenv==13.1.2
xlrd==1.1.0
xlwt==1.3.0
YURL==0.13
